Extract outline rendering to `outline` crate
