+++ crates/outline/Cargo.toml
@@ -19,6 +19,7 @@ gpui.workspace = true
 language.workspace = true
 ordered-float.workspace = true
 picker.workspace = true
+settings.workspace = true
 smol.workspace = true
 theme.workspace = true
 ui.workspace = true
