+++ crates/outline_panel/Cargo.toml
@@ -24,6 +24,7 @@ gpui.workspace = true
 language.workspace = true
 log.workspace = true
 menu.workspace = true
+outline.workspace = true
 project.workspace = true
 schemars.workspace = true
 search.workspace = true
